requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage==2.14.0
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==5.1.0
pyyaml==6.0.1
python-dateutil==2.8.2
